[FIX] delivery_roulier: Fix tests and add common
